Refactor helpers and update tests. Add pretty print helpers.
