+++ Tests/TinyMoonTests/TinyMoonTests.swift
@@ -3,7 +3,7 @@ import XCTest
 
 final class TinyMoonTests: XCTestCase {
   func test_tinyMoon_calculateMoonPhase_returnsFullMoon() throws {
-    let date = TestHelper.formatDate(year: 2024, month: 04, day: 23)
+    let date = MoonTestHelper.formatDate(year: 2024, month: 04, day: 23)
     let moon = TinyMoon.calculateMoonPhase(date)
 
     XCTAssertTrue(moon.isFullMoon())
@@ -13,7 +13,7 @@ final class TinyMoonTests: XCTestCase {
   }
 
   func test_moon_daysTillFullMoon_returnsCorrectDays() throws {
-    let date = TestHelper.formatDate(year: 2024, month: 04, day: 22)
+    let date = MoonTestHelper.formatDate(year: 2024, month: 04, day: 22)
     let moon = TinyMoon.calculateMoonPhase(date)
 
     XCTAssertFalse(moon.isFullMoon())
@@ -23,22 +23,22 @@ final class TinyMoonTests: XCTestCase {
   }
 
   func test_tinyMoon_calculateMoonPhase_returnsNewMoon() throws {
-    var date = TestHelper.formatDate(year: 2024, month: 11, day: 01)
+    var date = MoonTestHelper.formatDate(year: 2024, month: 11, day: 01)
     var moon = TinyMoon.calculateMoonPhase(date)
     XCTAssertEqual(moon.emoji, "\u{1F311}") // 🌑
     XCTAssertEqual(moon.daysTillNewMoon, 0)
 
-    date = TestHelper.formatDate(year: 2024, month: 12, day: 1)
+    date = MoonTestHelper.formatDate(year: 2024, month: 12, day: 1)
     moon = TinyMoon.calculateMoonPhase(date)
     XCTAssertEqual(moon.emoji, "\u{1F311}") // 🌑
     XCTAssertEqual(moon.daysTillNewMoon, 0)
   }
 
   func test_moon_uniquePhases() {
-    var emojisByMonth = [Helper.Month: Int]()
+    var emojisByMonth = [MonthTestHelper.Month: Int]()
 
-    Helper.Month.allCases.forEach { month in
-      let moons = TestHelper.moonMonth(month: month, year: 2024)
+    MonthTestHelper.Month.allCases.forEach { month in
+      let moons = MoonTestHelper.moonMonth(month: month, year: 2024)
       let emojis = moons.compactMap { moon in
         switch moon.moonPhase {
         case .newMoon, .firstQuarter, .fullMoon, .lastQuarter:
@@ -52,7 +52,7 @@ final class TinyMoonTests: XCTestCase {
       emojisByMonth[month] = emojis.count
     }
 
-    Helper.Month.allCases.forEach { month in
+    MonthTestHelper.Month.allCases.forEach { month in
       switch month {
       case .march, .december:
         XCTAssertEqual(emojisByMonth[month], 5)
